move QueryOptionsUtils
